refactor: replace LayerManager with Component base class
Major architectural change to UI component system:

- Add Component base class with AddChild() API for building UI trees
- Add MemoryArena for cache-friendly contiguous child storage
- Add zIndex to IComponent base for render ordering
- Add RenderContext::SetZIndex/GetZIndex for passing z-order through render pipeline
- Remove LayerManager (functionality moved to Component)
- Remove C++20 Layer/Focusable concepts (using interfaces instead)

Shapes now implement IComponent with zIndex support:
- Rectangle, Circle, Line, Text all support .zIndex in Args
- Shapes sorted by zIndex before rendering in Component::Render()

Updated all sandbox scenes to use new architecture.
Updated architecture.md documentation.

Note: Text stacking context bug identified - text renders after all
shapes due to separate batch renderers. Will be fixed in follow-up
uber shader epic.

+++ libs/ui/core/render_context.cpp
@@ -4,6 +4,6 @@
 namespace UI {
 
 	// Thread-local storage for current z-index during rendering
-	thread_local float RenderContext::s_currentZIndex = 0.0f;
+	thread_local short RenderContext::s_currentZIndex = 0;
 
 } // namespace UI
