+++ libs/ui/core/render_context.h
@@ -5,18 +5,18 @@ namespace UI {
 
 	/**
 	 * Thread-local rendering context
-	 * Used to pass z-index from LayerManager to shape Render() methods
+	 * Used to pass z-index from Component to shape Render() methods
 	 */
 	class RenderContext {
 	  public:
 		// Get the current rendering z-index
-		static float GetZIndex() { return s_currentZIndex; }
+		static short GetZIndex() { return s_currentZIndex; }
 
-		// Set the current rendering z-index (used by LayerManager)
-		static void SetZIndex(float zIndex) { s_currentZIndex = zIndex; }
+		// Set the current rendering z-index (used by Component)
+		static void SetZIndex(short zIndex) { s_currentZIndex = zIndex; }
 
 	  private:
-		static thread_local float s_currentZIndex;
+		static thread_local short s_currentZIndex;
 	};
 
 } // namespace UI
